in Domain.cpp: The method unID() returns a value of type ling int, now. Before, it was int.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/density-profiles@1066 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/Domain.h
@@ -353,7 +353,7 @@ public:
 	// prerequisite for the cylindrical coordinate system: R2max, centre, InvProfileUnit[3] => counterpart of setupProfile(), beyond that recordProfile() is extended
 	void sesDrop();
 	// number of the current bin a particle is located in. Actual parameters: global cartesian coordinates of the particle with respect to the simulation box coordinate system
-	int unID(double qx, double qy, double qz);
+	long int unID(double qx, double qy, double qz);
 	// returning the parameter that controls wheter or not a cylindrical profile is created. method called by Simulation::output()
 	bool isCylindrical();
 	// writing out a 3-dimensional density profile in cylindrical coordinates, counterpart of outputProfile.
